passthrough the lockfile in freeze function
