+++ micropip/_utils.py
@@ -1,10 +1,8 @@
 import functools
-import json
 from importlib.metadata import Distribution
 from pathlib import Path
 from sysconfig import get_config_var, get_platform
 
-from ._compat import LOCKFILE_PACKAGES
 from ._vendored.packaging.src.packaging.requirements import (
     InvalidRequirement,
     Requirement,
@@ -200,82 +198,6 @@ def check_compatible(filename: str) -> None:
         )
 
     raise ValueError(f"Wheel interpreter version '{tag.interpreter}' is not supported.")
-
-
-def fix_package_dependencies(
-    package_name: str, *, extras: list[str | None] | None = None
-) -> None:
-    """Check and fix the list of dependencies for this package
-
-    If you have manually installed a package and dependencies from wheels,
-    the dependencies will not be correctly setup in the package list
-    or the pyodide lockfile generated by freezing. This method checks
-    if the dependencies are correctly set in the package list and will
-    add missing dependencies.
-
-    Parameters
-    ----------
-    package_name (string):
-        The name of the package to check.
-
-    extras (list):
-        List of extras for this package.
-
-    """
-    if package_name in LOCKFILE_PACKAGES:
-        # don't check things that are in original repository
-        return
-
-    dist = Distribution.from_name(package_name)
-
-    package_requires = dist.requires
-    if package_requires is None:
-        # no dependencies - we're good to go
-        return
-
-    url = dist.read_text("PYODIDE_URL")
-
-    # If it wasn't installed with micropip / pyodide, then we
-    # can't do anything with it.
-    if url is None:
-        return
-
-    # Get current list of pyodide requirements
-    requires = dist.read_text("PYODIDE_REQUIRES")
-
-    if requires:
-        depends = json.loads(requires)
-    else:
-        depends = []
-
-    if extras is None:
-        extras = [None]
-    else:
-        extras = extras + [None]
-    for r in package_requires:
-        req = Requirement(r)
-        req_extras = req.extras
-        req_marker = req.marker
-        req_name = canonicalize_name(req.name)
-        needs_requirement = False
-        if req_marker is not None:
-            for e in extras:
-                if req_marker.evaluate(None if e is None else {"extra": e}):
-                    needs_requirement = True
-                    break
-        else:
-            needs_requirement = True
-
-        if needs_requirement:
-            fix_package_dependencies(req_name, extras=list(req_extras))
-
-            if req_name not in depends:
-                depends.append(req_name)
-
-    # write updated depends to PYODIDE_DEPENDS
-    (get_dist_info(dist) / "PYODIDE_REQUIRES").write_text(
-        json.dumps(sorted(x for x in depends))
-    )
 
 
 def validate_constraints(
